ch(auth): - implement authentication and authorization - configure context for state management
