+++ src/App.css
@@ -222,3 +222,69 @@ input[type="text"]:focus {
 .add-todo {
   cursor: pointer;
 }
+
+/* Login */
+.login {
+  background-color: #000;
+  display: flex;
+  flex-direction: column;
+  gap: 3em;
+  margin: 0 auto;
+  max-width: 450px;
+  height: 100vh;
+  padding: 5em 2em;
+}
+
+.login h1 {
+  font-size: 32px;
+  margin: 0;
+}
+
+.login form {
+  display: flex;
+  flex-direction: column;
+  gap: 1em;
+}
+
+.login input {
+  background-color: #1c1c1c;
+  padding: 25px 20px;
+  margin-bottom: 10px;
+  border: none;
+  color: #eee;
+  font-size: 14px;
+}
+
+.login button {
+  background-color: transparent;
+  border: 1px solid #eee;
+  color: #eee;
+  cursor: pointer;
+  padding: 25px 0;
+  font-family: inherit;
+  font-size: 16px;
+  margin-top: 40px;
+  text-transform: uppercase;
+  transition: background 300ms ease-in-out;
+}
+
+.login button:hover {
+  background-color: #eee;
+  color: #000;
+}
+
+.login input:focus {
+  outline: 1px solid #f5bcbc;
+}
+
+.login .error {
+  background-color: #d14a4a2d;
+  color: red;
+  display: inline-block;
+  padding: 3px 20px;
+  visibility: hidden;
+}
+
+.login .error.show {
+  visibility: visible;
+}
